+++ lib/m2m_keygen/request_signer/signed_request.rb
@@ -1,0 +1,10 @@
+# typed: strict
+
+module M2mKeygen
+  class RequestSigner
+    class SignedRequest < T::Struct
+      const :headers, T::Hash[String, String]
+      const :query, String
+    end
+  end
+end